rtkit: initial commit
Summary:
This is going to be used as part of pipewire (pulseaudio will also be updated to make use of this). This allows user-level pipewire/pulseaudio daemons to automatically nice themselves to a higher priority

NOTE that this depends on a kernel config change.
Depends on T9758

Test Plan: - Verified that pipewire had the appropriate nice value after a reboot with this enabled

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D11226